+++ 1-appeasing-the-masses.exs
@@ -1,0 +1,19 @@
+IO.puts("")
+IO.puts("# = = = = = = = = = = ")
+IO.puts("# Chapter 1")
+IO.puts("# Name: Appeasing the masses with code")
+IO.puts("# File: 1-appeasing-the-masses.exs")
+IO.puts("# = = = = = = = = = = ")
+IO.puts("")
+
+IO.puts("Exercise no. 1")
+result = 24 * 60
+IO.puts("Summary:")
+IO.puts(result)
+IO.puts("")
+
+IO.puts("Exercise no. 2")
+average = (4 + 8 + 15 + 16 + 23 + 42) / 6
+IO.puts("Average:")
+IO.puts(average)
+IO.puts("")