+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ## [0.0.2] (September 11 2022)
 
-**Updated :**
+**Added :**
 
-- README.md
+- 1-appeasing-the-masses.exs file
 
 ## [0.0.1] (March 11 2022)